Steve dev2 (#3)
* First Commit

Opened the scene and saved using version 5.6.1f1

* Ship Mouse Tracking

I have started to add the mouse tracking, the ship will now fly towards the mouse if the mouse is active. When the keyboard or joystick is used, the ship will then stop tracking the mouse. (Will require update to scene due to input checks)